% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellPhoneDB_convert.R
\name{CellPhoneDB_convert}
\alias{CellPhoneDB_convert}
\title{Convert CellPhoneDB database in a compatible format for ICELLNET analysis}
\usage{
CellPhoneDB_convert(
  complex = utils::read.csv(curl::curl(url =
    "https://raw.githubusercontent.com/ventolab/cellphonedb-data/master/data/complex_input.csv"),
    sep = ",", header = T, check.names = FALSE, stringsAsFactors = FALSE, na.strings =
    ""),
  interaction = utils::read.csv(curl::curl(url =
    "https://raw.githubusercontent.com/ventolab/cellphonedb-data/master/data/interaction_input.csv"),
    sep = ",", header = T, check.names = FALSE, stringsAsFactors = FALSE, na.strings =
    ""),
  gene_info = utils::read.csv(curl::curl(url =
    "https://raw.githubusercontent.com/ventolab/cellphonedb-data/master/data/gene_input.csv"),
    sep = ",", header = T, check.names = FALSE, stringsAsFactors = FALSE, na.strings =
    ""),
  ppi = T
)
}
\arguments{
\item{complex}{complex_input.csv file from Github (ventolab/cellphonedb-data)}

\item{interaction}{interaction_input.csv file from Github (ventolab/cellphonedb-data)}

\item{gene_info}{gene_input.csv file from Github (ventolab/cellphonedb-data)}

\item{ppi}{logical, to retain only ppi interactions. TRUE by default}
}
\description{
Provide CellPhoneDB ligand-receptor database in a format suitable for ICELLNET analysis
}
\examples{
\dontrun{CellPhoneDB_convert()}

}
