% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.hgu133plus2.R
\name{db.hgu133plus2}
\alias{db.hgu133plus2}
\title{Conversion of affymetrix probe name into gene symbol}
\usage{
db.hgu133plus2(db, data)
}
\arguments{
\item{db}{Ligand/receptor database}

\item{data}{List of annotation data (ex: rownames of the transcriptomic profiles matrix)}
}
\description{
This function converts LR database gene symbol annotation into Affymetrix Human Genome U133 Plus 2.0 Array annotation.
}
\details{
This function transform the annotation of the database into probeset annotation, only available for Affymetrix Human Genome U133 Plus 2.0 Array.
}
\examples{
\dontrun{
db.hgu133plus2(db,data.probes)
}


}
