% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_edges.R
\name{geom_edges}
\alias{geom_edges}
\title{Display the edges in the ICELLNET network.}
\usage{
geom_edges(
  mapping = NULL,
  data = NULL,
  position = "identity",
  arrow = NULL,
  curvature = 0,
  angle = 90,
  ncp = 5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{arrow}{arrow}

\item{curvature}{curvature}

\item{angle}{angle}

\item{ncp}{ncp}

\item{na.rm}{na.rm}

\item{show.legend}{to show the legend or not}

\item{inherit.aes}{inherit.aes}
}
\description{
Display the edges in the ICELLNET network.
}
\examples{
\dontrun{geom_edges (data = my_df_net_2, aes(size = Score), arrow = arrow( length = unit(7, "pt"), angle = 35, type = "open", ends = "first") }

}
