% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_node_label.R
\name{geom_node_label}
\alias{geom_node_label}
\title{Label the node with the name of the cell type in the ICELLNET network.}
\usage{
geom_node_label(mapping = NULL, data = NULL, position = "identity",
  ..., parse = FALSE, nudge_x = 0, nudge_y = 0,
  label.padding = unit(0.25, "lines"), label.r = unit(0.15, "lines"),
  label.size = 0.25, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{parse}{parse}

\item{nudge_x}{nudge_x}

\item{nudge_y}{nudge_y}

\item{label.padding}{label.padding}

\item{label.r}{label.r}

\item{label.size}{label.size}

\item{na.rm}{na.rm}

\item{show.legend}{to show the legend or not}

\item{inherit.aes}{inherit.aes}
}
\description{
Label the node with the name of the cell type in the ICELLNET network.
}
\examples{
\dontrun{geom_node_label(aes(label = Cell_type, fill = Cell_type), size = rel(6), fontface = "bold") }
}
