% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nodes.R
\name{geom_nodes}
\alias{geom_nodes}
\title{Display the nodes in the ICELLNET network.}
\usage{
geom_nodes(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{to show the legend or not}

\item{inherit.aes}{inherit.aes}
}
\description{
Display the nodes in the ICELLNET network.
}
\examples{
\dontrun{geom_nodes(aes(color = Cell_type), size = 30) }
}
