\name{icellnet-package}
\alias{icellnet-package}
\alias{icellnet}
\docType{package}
\title{
\packageTitle{icellnet}
}
\description{
\packageDescription{icellnet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{icellnet}
\packageIndices{icellnet}
}
\author{
\packageAuthor{icellnet}

Maintainer: \packageMaintainer{icellnet}
}
\keyword{ package }
