% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icellnet.ind.score.R
\name{icellnet.ind.score}
\alias{icellnet.ind.score}
\title{ICELLNET score computation for replicates, to compute the ovalue (not exported)}
\usage{
icellnet.ind.score(
  direction = c("out", "in"),
  CC.data = CC.data,
  PC.data = PC.data,
  PC = PC,
  PC.target = PC.target,
  cell = cell,
  CC.type = c("RNAseq", "Microarray"),
  PC.type = c("RNAseq", "Microarray"),
  db = db,
  between = c("cells", "conditions")
)
}
\arguments{
\item{direction}{Direction of the communication (either "out" or "in")}

\item{CC.data}{Data.frame of transcriptomic profiles corresponding to the central cell}

\item{PC.data}{Data.frame of transcriptomic profiles corresponding to the partner cells}

\item{PC}{Vector selecting a list of partnercell for the cell-cell communication analysis}

\item{PC.target}{Data.frame of information concerning the PC.data samples}

\item{cell}{A cell type}

\item{CC.type}{Type of transcriptomic data for the central cell (either "RNAseq" or "Microarray")}

\item{PC.type}{Type of transcriptomic data for the partner (either "RNAseq" or "Microarray")}

\item{db}{Database of ligand/receptors interactions}

\item{between}{Type of pvalue matrix desired : "cells" to compare communication scores of the central cell with different partner cells, "conditions" to compare the communication score of two central cells in different biological conditions with the same partner cell.}
}
\description{
Needed to compute the pvalue in the icellnet.score.pvalue function. Returns communication scores between a central cell (CC) and partner cell types (PC).
}
\details{
See icellnet.score.pvalue function for more details.

The function returns the communication score per replicates.
}
\examples{
\dontrun{icellnet.ind.score ( direction = "in", CC.data = CC.data, PC.data = PC.data,...
PC = PC, PC.target = PC.target, cell=cell, CC.type = "RNAseq", PC.type="RNAseq" , db = db, ...
between="cells")}
}
