% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icellnet.score.pvalue.R
\name{icellnet.score.pvalue}
\alias{icellnet.score.pvalue}
\title{Comparison of different communication score by computing a pvalue matrix}
\usage{
icellnet.score.pvalue(
  direction = c("out", "in"),
  CC.data = CC.data,
  CC.data2 = NULL,
  PC.data = PC.data,
  PC = PC,
  PC.target = PC.target,
  CC.type = c("RNAseq", "Microarray"),
  PC.type = c("RNAseq", "Microarray"),
  db = db,
  between = c("cells", "conditions"),
  method = "BH"
)
}
\arguments{
\item{direction}{Direction of the communication (either "out" or "in"). "out" represent the communication from the central cell to the peripheral cells, which means that the ligands considered are expressed by the central cell, and the receptors are expressed by the peripheral cells. "in" represents the communication from the peripheral cells to the central cells, which means that the considered ligands are expressed by the peripheral cells and the receptors are expressed by the central cells.}

\item{CC.data}{Data.frame of transcriptomic profiles corresponding to the central cell}

\item{CC.data2}{Data.frame of transcriptomic profiles corresponding to the central cell in a different biological condition. NULL as a default.}

\item{PC.data}{Data.frame of transcriptomic profiles corresponding to the peripheral cells}

\item{PC}{Vector selecting a list of peripheral cell for the cell-cell communication analysis}

\item{PC.target}{Data.frame of information concerning the PC.data samples}

\item{CC.type}{Type of transcriptomic data for the central cell (either "RNAseq" or "Microarray")}

\item{PC.type}{Type of transcriptomic data for the peripheral (either "RNAseq" or "Microarray")}

\item{db}{Database of ligand/receptors interactions}

\item{between}{Type of pvalue matrix desired : "cells" to compare communication scores of the central cell with different peripheral cells, "conditions" to compare the communication score of two central cells in different biological conditions with the same peripheral cell.}

\item{method}{method to adjust the pvalue. See p.adjust() function for the available methods. "BH" method is set as a default.}
}
\description{
Return a pvalue matrix to allow communication score comparison.
}
\details{
Two types of pvalue can be computed, to compare either the communication scores obtained from the same central cell to different peripheral cells (between="cells"),
or to compare communication scores obtained from two different central cells corresponding to different biological conditions with the same peripheral cell (between="conditions").
If between="cells", the communication score is computed considering the average expression of ligands for the central cell, and each replicates separately
for the receptor expression of the peripheral cells. In this way, for one peripheral cell type, we obtain a distribution of n communication scores, n beeing the number
of peripheral cells replicates for this particular cell type. If between="conditions", then, the communication score is computed considering each replicates of the central
cell separately, and the average gene expression for the peripheral cells. We obtain a distribution of n communication scores, n beeing the number
of central cell replicates in one biological condition.
Then, a Wilcoxon statistical test is performed to compare the communication scores distributions. The pvalues are ajusted with p.adjust(), with "BH" method as a default.
It returns the pvalue matrix of statistical tests, that can be visualize in a heatmap with the pvalue.plot function.
}
\examples{
\dontrun{
icellnet.score.pvalue( CC.data= CC.data.selection.S1, PC.data=PC.data, PC = my.selection,...
 PC.target = PC.target, CC.type = "RNAseq", PC.type = "Microarray", direction="out", between="cells")
 }
}
