% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icellnet.score.R
\name{icellnet.score}
\alias{icellnet.score}
\title{Communication score computation with ICELLNET method}
\usage{
icellnet.score(direction = c("out", "in"), CC.data = CC.data,
  PC.data = PC.data, PC = PC, PC.target = PC.target,
  CC.type = c("RNAseq", "Microarray"), PC.type = c("RNAseq",
  "Microarray"), db = db, family.type = "Family")
}
\arguments{
\item{direction}{Direction of the communication (either "out" or "in")}

\item{CC.data}{Data.frame of transcriptomic profiles corresponding to the central cell}

\item{PC.data}{Data.frame of transcriptomic profiles corresponding to the peripheral cells}

\item{PC}{Vector selecting a list of peripheral cell for the cell-cell communication analysis}

\item{PC.target}{Data.frame of information concerning the PC.data samples}

\item{CC.type}{Type of transcriptomic data for the central cell (either "RNAseq" or "Microarray")}

\item{PC.type}{Type of transcriptomic data for the peripheral (either "RNAseq" or "Microarray")}

\item{db}{Database of ligand/receptors interactions}

\item{family.type}{"Family" as a default, to separate the different family of molecules in the database. Can be set to "Subfamily" to consider the subfamilies of cytokines.}
}
\description{
Returns global communication scores between a central cell (CC) and peripheral cell types (PC).
}
\details{
From the CC and PC expression matrices (RNA-seq or Microarray), this function computes a communication score using the knowledge-based LR database.
It selects ligands and receptors from the database an expressed by the cells of interest. A selection of the PC can be done.
The direction of the communication is indicated using the "direction" argument:
- out == compute the communication score from the CC to PC
- in == compute the communication score from the PC to CC

The function returns the global communication score and the individual communication score matrices.
}
\examples{
\dontrun{icellnet.score( CC.data= CC.data.selection.S1, PC.data=PC.data, PC = my.selection, ...
PC.target = PC.target, CC.type = "RNAseq", PC.type = "Microarray", direction="out")}

}
