% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand.average.MA.R
\name{ligand.average.MA}
\alias{ligand.average.MA}
\title{ligand.average.MA function}
\usage{
ligand.average.MA(db = db, data = data, SYMBOL = rownames(data))
}
\arguments{
\item{db}{Ligand/receptor database}

\item{data}{A matrix of transcriptomic profiles}

\item{SYMBOL}{List of gene symbols}
}
\description{
Compute the average expression for all ligands expressed in dataset provided, only available for Affymetrix Human Genome U133 Plus 2.0 Array.
}
\details{
This function returns the mean of expression of each ligand (included in the database db) of several transcriptomic profiles that should
correspond to the same biological condition. If the ligand is composed of several subunits, the function computes the geometric mean of expression
of the mean values obtained for each subunits.
}
\examples{
\dontrun{ligand.average.MA( db=db, data = data,  SYMBOL = rownames(data))}
}
