% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand.average.R
\name{ligand.average}
\alias{ligand.average}
\title{ligand.average function}
\usage{
ligand.average(db = db, data = data)
}
\arguments{
\item{db}{Ligand/receptor database}

\item{data}{A dataframe of transcriptomic profiles with gene names as rownames}
}
\description{
Computes the average expression for all ligands expressed in the dataset provided with gene symbol annotation.
}
\details{
This function returns the mean of expression of each ligand (included in the database db) of several transcriptomic profiles that should
correspond to the same biological condition. If the ligand is composed of several subunits, the function computes the geometric mean of expression
of the mean values obtained for each subunits.
}
\examples{
\dontrun{ ligand.average(db=db, data = data)}
}
