% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.family.barplot.R
\name{LR.family.barplot}
\alias{LR.family.barplot}
\title{Barplot of family of molecules contribution to the communication scores}
\usage{
LR.family.barplot(table = table, title = NULL, ymax = NULL)
}
\arguments{
\item{table}{Matrix of the contribution per family of molecules to the  global communication score (sum on each column should be equal to the global communication score)}

\item{title}{Title of the barplot}

\item{ymax}{Limits of the yaxis. If ymax=NULL, ymax takes the maximum value of the table.}
}
\description{
Display the contribution of families or subfamilies of molecules (included in my.selection) to the global communication scores.
The matrix taken as an argument should be results of the LR.family.score() function.
}
\examples{
\dontrun{
LR.family.barplot(LR.family.score)
}

}
