% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.heatmap.R
\name{LR.heatmap}
\alias{LR.heatmap}
\title{Plot of individual communication scores of contributing L/R pairs.}
\usage{
LR.heatmap(
  lr = lr,
  thresh = 0,
  topn = NULL,
  sort.by = "sum",
  db.couple = NULL,
  title = title,
  value_display = NA
)
}
\arguments{
\item{lr}{Matrix of individual communication scores}

\item{thresh}{Value set as a threshold to display the L/R pairs contributing with a score superior to this threshold}

\item{topn}{Value set as top n interactions to display}

\item{sort.by}{character, "sum" or "var", "sum" as default. In combination with topn parameter, allows to sort L/R pairs according to the most contributing ("sum"), or the most different among conditions ("var")}

\item{db.couple}{output of the name.lr.couple() function. name.lr.couple(db, "Family") is used as default}

\item{title}{Title of the heatmap}

\item{value_display}{Numerical value, used as a threshold value to add numerical values to the heatmap above this value.}
}
\description{
Display the individual communication scores of ligand/receptor pairs that contributes with a score superior to a threshold and/or the top n interactions contributing to the score. Return a ggplot object.
}
\examples{
\dontrun{LR.heatmap(lr = lr1, topn = 30, sort.by = "sum")
}

}
