% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.viz.R
\name{LR.viz}
\alias{LR.viz}
\title{Plot individual communication scores of a specific L/R interactions for all cell pairs.}
\usage{
LR.viz(data = data, couple = couple, db = db, plot = TRUE)
}
\arguments{
\item{data}{A dataframe of transcriptomic profiles with gene names as rownames}

\item{couple}{character, name of a interaction "L / R". If several subunits, refer to output of name.lr.couple() function for interaction names.}

\item{db}{Ligand/receptor database}

\item{plot}{logical FALSE/TRUE (TRUE by default). If TRUE, the function returns the graphical representation of communication score for each cell pairs as a heatmap. If FALSE, the communication score matrix is returned.}
}
\description{
Plot individual communication scores of a specific L/R interactions for all cell pairs. Useful to see specificity of an interaction between two cell types. Relevant only for single cell datasets, deciphering communication between clusters. Returns either a ggplot object (if plot=TRUE) or a matrix.
}
\examples{
\dontrun{LR.viz(data=data.icell , couple=couple, db=db , plot=TRUE)}

}
