% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.lr.couple.R
\name{name.lr.couple}
\alias{name.lr.couple}
\title{name.lr.couple function}
\usage{
name.lr.couple(db = db, type = c("Family", "Subfamily"))
}
\arguments{
\item{db}{Ligand/receptor database}

\item{type}{Can be set to "Family" or "Subfamily", to display the  family/subfamily of molecules which the L/R interaction belongs to.}
}
\description{
From the database of ligand/receptors interactions, the function retrieves the corresponding gene symbols.
It returns a table containing the name of the couple L/R and the family/subfamily of molecules which the interaction belongs to.
}
\examples{
name.lr.couple(db = db, type="Family")
}
