% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.create.R
\name{network.create}
\alias{network.create}
\title{Display the communication network with ICELLNET method}
\usage{
network.create(
  icn.score,
  scale = c(min(icn.score), max(icn.score)),
  direction = c("in", "out"),
  PC.col = NULL
)
}
\arguments{
\item{icn.score}{Matrix of global communication scores}

\item{scale}{vector that contains the minimum and the maximum of the global communication score matrix.}

\item{direction}{Direction of the communication (either "out" or "in")}

\item{PC.col}{Color vector for the partner cells c("cell1"= "color1", "cell2"="color2").}
}
\description{
Display the network from global communication scores matrix between a central cell (CC) and partner cell types (PC).
}
\details{
The matrix of global communication score is obtained with the icellnet.score function.
- out == display the communication score from the CC to PC
- in == display the communication score from the PC to CC
If you want to compare the communication between different conditions, we suggest to bind the different global
communication score matrix into a single one, so that all the scores are scaled together to display the
network.
}
\examples{
\dontrun{network.create(icn.score = Scores.norm, scale = c(round(min(Scores.norm)),...
round(max(Scores.norm))), direction = "out", PC.col)}

}
