% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.data.all.R
\docType{data}
\name{PC.data.all}
\alias{PC.data.all}
\title{PC.data.all}
\format{A data frame with 53675 rows corresponding to the gene expression and 746 columns corresponding to the different samples.}
\source{
\url{http://biogps.org/dataset/species/human/}
}
\usage{
PC.data.all
}
\description{
Dataset integrating transcriptional profiles of over twenty human primary cell types generated with Affymetrix Human Genome U133 Plus 2.0 Array.
}
\keyword{datasets}
