% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.target.all.R
\docType{data}
\name{PC.target.all}
\alias{PC.target.all}
\title{PC.target.all}
\format{A data frame with 53675 rows corresponding to the gene expression and 746 columns corresponding to the differents samples. All these samples have ben generated
using Affymetrix Human Genome U133 Plus 2.0 Array technology. dataset already processed considering all the samples.

' \describe{
  \item{ID}{Complete name of the sample in BioGPS}
  \item{Sample}{Name of the sample in BioGPS}
  \item{Dataset_description}{Description of the study}
  \item{Chip_ID}{Chip ID}
  \item{Series}{GSE number}
  \item{Chip_description}{Description}
  \item{PMID}{Pubmed ID}
  \item{Cell_description}{Cell description}
  \item{Cell_type}{Cell type, first classification}
  \item{Class}{Cell type classification that is used in Icellnet package. you can modify this column by the Cell_type column if you prefer this other classification, less detailed.}
}}
\source{
\url{http://biogps.org/dataset/species/human/}
}
\usage{
PC.target.all
}
\description{
Dataset describing the samples of PC.data.all
}
\keyword{datasets}
