% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perc_exp_infos.R
\name{Perc_exp_infos}
\alias{Perc_exp_infos}
\title{Perc_exp_infos function}
\usage{
Perc_exp_infos(object = object, gene = gene, cell_id = cell_id)
}
\arguments{
\item{object}{SeuratObject with clusters as Idents(), and with slot counts and data.}

\item{gene}{Character, name of the gene of interest}

\item{cell_id}{Name/identity of the cluster, found also in Idents(object).}
}
\description{
Compute the average expression for a gene of interest for one cluster (cell_id) and the percentage of cells expressing the gene in the cluster.
}
\details{
Starting from a SeuratObject, this function returns as a vector: 1) the percentage of cells expressing the gene (counts >0), 2) the average gene expression for the cluster (slot=data from the SeuratObject)
}
\examples{
\dontrun{Perc_exp_infos(object=object, gene=gene, cell_id=cell_id)}

}
