% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.plot.R
\name{pvalue.plot}
\alias{pvalue.plot}
\title{Comparison of different communication score by displaying pvalues as a correlation plot}
\usage{
pvalue.plot(pvalue, PC = PC)
}
\arguments{
\item{pvalue}{pvalue matrix. See icellnet.score.pvalue() function for more information}

\item{PC}{Vector selecting a list of peripheral cell for the cell-cell communication analysis}
}
\description{
Display pvalue matrix to compare different communication scores between the central cell and different peripheral cells
}
\examples{
\dontrun{
pvalue.matrix <- icellnet.score.pvalue(...)
pvalue.plot(pvalue=pvalue.matrix, PC=PC)
}


}
