% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor.average.MA.R
\name{receptor.average.MA}
\alias{receptor.average.MA}
\title{receptor.average.MA function}
\usage{
receptor.average.MA(db = db, data = data, SYMBOL = rownames(data))
}
\arguments{
\item{db}{Ligand/receptor database}

\item{data}{A matrix of transcriptomic profiles}

\item{SYMBOL}{List of gene symbols}
}
\description{
Compute the average expression for all receptors expressed in dataset provided, only available for Affymetrix Human Genome U133 Plus 2.0 Array.
}
\details{
This function returns the mean of expression of each receptor (included in the database db) of several transcriptomic profiles that should
correspond to the same biological condition. If the receptor is composed of several subunits, the function computes the geometric mean of expression
of the mean values obtained for each subunits.
}
\examples{
\dontrun{receptor.average.MA ( db=db, data = data,  SYMBOL = rownames(data))}

}
