% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor.average.R
\name{receptor.average}
\alias{receptor.average}
\title{receptor.average function}
\usage{
receptor.average(db = db, data = data)
}
\arguments{
\item{db}{Ligand/receptor database}

\item{data}{A dataframe of transcriptomic profiles with gene names as rownames}
}
\description{
Computes the average expression for all receptors expressed in the dataset provided with gene symbol annotation.
}
\details{
This function returns the mean of expression of each receptor (included in the database db) of several transcriptomic profiles that should
correspond to the same biological condition. If the receptor is composed of several subunits, the function computes the geometric mean of expression
of the mean values obtained for each subunits.
}
\examples{
\dontrun{receptor.average(db=db, data = data)}

}
