% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.data.cleaning.R
\name{sc.data.cleaning}
\alias{sc.data.cleaning}
\title{Filtering scRNAseq data and computing average expression per cluster}
\usage{
sc.data.cleaning(
  object = object,
  assay = "RNA",
  db = db,
  filter.perc = NULL,
  save_file = T,
  path = NULL,
  force.file = FALSE
)
}
\arguments{
\item{object}{Seurat object with cells category (as future partner cells as Idents())}

\item{assay}{Seurat assay, RNA assay by default.}

\item{db}{Ligand/receptor database}

\item{filter.perc}{Number between 0 and 100 (percentage), used to filter gene expression in each cluster (see details)}

\item{save_file}{Logical, TRUE as default. If TRUE, it will save the file scRNAseq_statsInfo_for_ICELLNET.csv, that will be reused automatically if ever needed.}

\item{path}{path to save the file scRNAseq_statsInfo_for_ICELLNET.csv. Working directory used as default path.}

\item{force.file}{Logical, FALSE as default. If scRNAseq_statsInfo_for_ICELLNET.csv already exists, set force.file=T will regenerate automatically the intermediate statistics table for data filtering.}
}
\description{
This function computes average expression per cluster from a Seurat object and eventually filters genes to keep only genes expressed by a specific percentage of the cells in their respective cluster. It returns the average the gene expression matrix (filtered) with clusters in columns.
}
\details{
This function starts from a Seurat object and will used object[['RNA']]@data (as default, other it has to be specified) to compute average expression for each gene and for each cluster.
The function will also generate a statistic file (csv format) displaying for each gene in each cluster :
1) the percentage of cells expressing the gene for this specific cluster
2) the gene average expression value, considering all cells from the cluster.
If this file is already computed by a previous use of the function, it will be used automatically. To regenerate the file agan, set force.file=TRUE.
If filter.perc is not NULL and set to 5, the gene expression matrix will be filtered to keep only genes expressed by at least 5% of the cells at the cluster level. Other genes (filtered) will be set at 0 for the corresponding cluster(s).
}
\examples{
\dontrun{
sc.data.cleaning(object=seurat, db = icellnet_db, filter.perc=NULL, save_file=T, path=NULL)
}

}
