% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icellnet.ind.score.R
\name{icellnet.ind.score}
\alias{icellnet.ind.score}
\title{ICELLNET score computation for replicates, to compute the ovalue (not exported)}
\usage{
icellnet.ind.score(direction = c("out", "in"), CC.data = CC.data,
  PC.data = PC.data, PC = PC, PC.target = PC.target, cell = cell,
  CC.type = c("RNAseq", "Microarray"), PC.type = c("RNAseq",
  "Microarray"), db = db, family.type = "Family",
  between = c("cells", "conditions"))
}
\arguments{
\item{direction}{Direction of the communication (either "out" or "in")}

\item{CC.data}{Data.frame of transcriptomic profiles corresponding to the central cell}

\item{PC.data}{Data.frame of transcriptomic profiles corresponding to the peripheral cells}

\item{PC}{Vector selecting a list of peripheral cell for the cell-cell communication analysis}

\item{PC.target}{Data.frame of information concerning the PC.data samples}

\item{cell}{A cell type}

\item{CC.type}{Type of transcriptomic data for the central cell (either "RNAseq" or "Microarray")}

\item{PC.type}{Type of transcriptomic data for the peripheral (either "RNAseq" or "Microarray")}

\item{db}{Database of ligand/receptors interactions}

\item{family.type}{"Family" as a default, to separate the different family of molecules in the database. Can be set to "Subfamily" to consider the subfamilies of cytokines.}

\item{between}{Type of pvalue matrix desired : "cells" to compare communication scores of the central cell with different peripheral cells, "conditions" to compare the communication score of two central cells in different biological conditions with the same peripheral cell.}
}
\description{
Needed to compute the pvalue in the icellnet.score.pvalue function. Returns communication scores between a central cell (CC) and peripheral cell types (PC).
}
\details{
See icellnet.score.pvalue function for more details.

The function returns the communication score per replicates.
}
\examples{
\dontrun{icellnet.ind.score ( direction = "in", CC.data = CC.data, PC.data = PC.data,...
PC = PC, PC.target = PC.target, cell=cell, CC.type = "RNAseq", PC.type="RNAseq" , db = db, ...
family.type = "Family", between="cells")}
}
