% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.balloon.plot.R
\name{LR.balloon.plot}
\alias{LR.balloon.plot}
\title{Plot of individual communication scores of contributing L/R pairs.}
\usage{
LR.balloon.plot(lr = lr, PC = PC, thresh = thresh, type = c("raw",
  "percentage"), db.name.couple = db.name.couple, title = title)
}
\arguments{
\item{lr}{Matrix of individual communication scores}

\item{PC}{Vector of selected peripheral cell types}

\item{thresh}{Value set as a threshold to display the L/R pairs contributing with a score superior to this threshold}

\item{type}{"raw" as default, can be set to "percentage" if you wish to consider the threshold as a defined percentage of the global communication score}

\item{db.name.couple}{output of the name.lr.couple() function. name.lr.couple(db, "Family") is set as a default}

\item{title}{Title of the balloon plot}
}
\description{
Display the individual communication scores of ligand/receptor pairs that contributes with a score superior to a threshold (thresh = 10 as a default)
}
\examples{
\dontrun{lr <- icellnet.score.results[[2]]
LR.balloon.plot(lr = lr, PC =PC , thresh = 20,  db.name.couple=name.lr.couple(db, "Family"))
}

}
