% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.family.score.R
\name{LR.family.score}
\alias{LR.family.score}
\title{Distribution of class of molecules contribution to the communication scores}
\usage{
LR.family.score(
  lr = lr,
  my.family = my.family,
  db.couple = as.data.frame(db.couple)
)
}
\arguments{
\item{lr}{Matrix of individual communication scores (sum on each column should be equal to the global communication score)}

\item{my.family}{Vector selecting family or subfamilies of molecules for the analysis}

\item{db.couple}{Output of the name.lr.couple() function. name.lr.couple(db, "Family") is set as a default}
}
\description{
Computes the contribution of families or subfamilies of molecules (included in my.selection) to the global communication scores.
The function returns a table that can be visualized in a barplot using the LR.family.barplot() function.
}
\examples{
\dontrun{LR.family.score(lr=lr, my.family=my.family, db.couple=db.name.couple)}

}
